requirements.txt
selenium == 4.6.0
pytest == 7.2.1
allure-pytest==2.12.0
pytest-xdist==2.5.0
Appium-Python-Client==2.11.1